google-cloud-ai-certifications/professional-ml-engineer/code-examples/requirements.txt
google-cloud-aiplatform>=1.71.1
google-cloud-bigquery>=3.25.0
tensorflow>=2.17.0
tensorflow-datasets>=4.9.6
tensorflow-serving-api>=2.17.0
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
mlflow>=2.16.0
google-cloud-storage>=2.18.0
google-cloud-pipeline-components>=2.16.0
protobuf>=4.25.0
